requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-heroku==0.3.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
